Fix PR21996 by updating DEREF_FAULT for dyninst.
* runtime/loc2c-runtime.h (DEREF_FAULT): Fix dyninst compile errors on
  32-bit systems where 'intptr_t' is an 'int'. Be sure to cast to 'long'
  in DEREF_FAULT and STORE_DEREF_FAULT before printing the value with
  "%lx".

+++ runtime/loc2c-runtime.h
@@ -229,14 +229,14 @@
 #else
 #define DEREF_FAULT(addr) ({						    \
     snprintf(c->error_buffer, sizeof(c->error_buffer),			    \
-      STAP_MSG_LOC2C_01, (intptr_t)(addr));   \
+	     STAP_MSG_LOC2C_01, (long)(intptr_t)(addr));		    \
     c->last_error = c->error_buffer;					    \
     goto deref_fault;							    \
     })
 
 #define STORE_DEREF_FAULT(addr) ({					    \
     snprintf(c->error_buffer, sizeof(c->error_buffer),			    \
-      STAP_MSG_LOC2C_02, (intptr_t)(addr));  \
+	     STAP_MSG_LOC2C_02, (long)(intptr_t)(addr));		    \
     c->last_error = c->error_buffer;					    \
     goto deref_fault;							    \
     })
